[Operator] move ops weight_norm/instance_norm/outer/celoss into fused directory, which are registered as AutogradCUDA before
