+++ src/flag_gems/__init__.py
@@ -155,10 +155,8 @@ def enable(
                 weight_norm_interface_backward,
                 Autograd.disable,
             ),
-            ("_weight_norm", weight_norm, Autograd.enable),
             ("gt.Tensor", gt, Autograd.disable),
             ("gt.Scalar", gt_scalar, Autograd.disable),
-            ("instance_norm", instance_norm, Autograd.enable),
             ("isfinite", isfinite, Autograd.disable),
             ("isin.Tensor_Tensor", isin, Autograd.disable),
             ("isin.Scalar_Tensor", isin, Autograd.disable),
@@ -266,8 +264,6 @@ def enable(
             ("quantile", quantile, Autograd.disable),
             ("_log_softmax", log_softmax, Autograd.disable),
             ("_log_softmax_backward_data", log_softmax_backward, Autograd.disable),
-            ("outer", outer, Autograd.enable),
-            ("cross_entropy_loss", cross_entropy_loss, Autograd.enable),
             ("nll_loss_forward", nll_loss_forward, Autograd.disable),
             ("nll_loss_backward", nll_loss_backward, Autograd.disable),
             ("nll_loss2d_forward", nll_loss2d_forward, Autograd.disable),
