+++ src/flag_gems/fused/__init__.py
@@ -1,8 +1,12 @@
+from .cross_entropy_loss import cross_entropy_loss
 from .gelu_and_mul import gelu_and_mul
+from .instance_norm import instance_norm
+from .outer import outer
 from .rotary_embedding import apply_rotary_pos_emb
 from .silu_and_mul import silu_and_mul
 from .skip_layernorm import skip_layer_norm
 from .skip_rms_norm import skip_rms_norm
+from .weight_norm import weight_norm
 
 __all__ = [
     "apply_rotary_pos_emb",
@@ -10,4 +14,8 @@ __all__ = [
     "skip_rms_norm",
     "silu_and_mul",
     "gelu_and_mul",
+    "cross_entropy_loss",
+    "outer",
+    "instance_norm",
+    "weight_norm",
 ]
